SwfBlockPointer: Change pointer size to u32.
It needs to be u32 or it won't be possible to read the last
byte of the block, u16 isn't big enough.

+++ src/file/block_pointer.rs
@@ -1,3 +1,3 @@
 /// A pointer to a specific byte within a [SwfBlock][super::block::SwfBlock].
 #[derive(Clone, Copy, PartialEq, PartialOrd, Eq, Ord, Debug, Hash)]
-pub struct SwfBlockPointer(pub(super) u16);
+pub struct SwfBlockPointer(pub(super) u32);
